[#2131] Handle capability modes when declaring capabilities
Problem: When we declare our capabilities, we don't take the
`capability_mode` into consideration. E.g. even if only semantic tokens
are enabled, we'd say we support document symbols. This causes some
problems with some plugins, like `nvim_navic`.

Solution: In `is_request_enabled`, if we are not dealing with
(supported) semantic tokens, or we are dealing with semantic tokens but
it's disabled, then say it's not supported. Otherwise, continue with the
old logic.

Also, I formatted some unformatted code.