[#2134] Fix `data` variable usage
Problem: typechecker says "variable `data` may be not initialized".

It may remain uninitialized if exception strikes (`ligo` executable is
not found).

Solution: move all the following logic that uses `data` variable into
`try` block and make it part of separate
function (`updateLigoUnchecked`).

This slightly changes the logic: previously, if the proposed LIGO
install/upgrade was applied, we would use `data` variable to verify
whether the version is good. But there seems to be no way in doing that
- this way we check the version of the old `ligo` which is just getting
upgraded.