[#2134] Avoid `activate` function in extensions
Problem: typechecker reports a warning about the class members, that
they are not initialized with constructor. We initialize them in
`activate` method, so there is really a period of time when those
class members remain uninitialized.

Solution: turn `activate` into a constructor.

Now `LigoExtension` abstract class is mostly redundant, so I also
replace it with the use of `Disposable`.