[#2134] Better dispose of DebuggerExtension
Problem: `DebuggerExtension` class exposes `server`, but this is not
what we want. `server` comes via a constructor, and is used by
`LspExtension` too, so making debugger extension responsible for
disposing it is wrong.

Solution: dispose `server` in `extension.ts`, not inside
`DebuggerExtension`.

Also, `DebuggerExtension` has to dispose something after all - it is
step granularity status button, so do this inside `dispose()`.