[#1751] Improve error-recovery until all tests pass
Problem: Many tests are not passing after all the changes to the typer.
It would be nice to bring them all to a state where they pass to
continue.

Solution: Most changes involve changing the typer until a lot of places
where `raise.error` are called are eliminated. For this, I created an
`Error_recovery` module that will have some useful helpers for us. With
this, many places that throw exceptions are instead replaced with an
equivalent that will allow recovering with a dummy value. There are
still places that may throw, but at least many were eliminated.

I ceated a new flag to `ligo info get-scope`, `ligo print ast-typed` and
`ligo print ast-aggregated`, called `--typer-error-recovery`. This is
mostly for debugging, and it's a bit hacky since it sets
`fast_fail:false`, but it was useful specially to find bugs out.

I promoted many tests. It's a bit annoying that now there are
underspecified type errors, but at least all the others are still there.

`Get_scope` was changed to make better use of `Trace`, so now it should
be able to display more diagnostics.

Types pass was changed to appropriately collect errors, and initialize
the environment (which was previously erroneously removed).

The typer now logs the errors with `raise.log_error`, so I no longer
store them in the AST. Less `E_error`s should be produced now, but in
case they do (because `raise.error` was called), then they will be
caught, logged with `raise.log_error`, and an `E_error` will be
produced.