[#1751] Improve printing of generated variables
Problem: We see `gen#N` variables which is bad, so we should improve
those names.

Solution: Export the table of `make_concise_name` in order to replace
them in error messages. Unfortunately, there are some duplicated
diagnostics now, so I have to `dedup_and_sort` them.

I had a leftover in the untyper, where I always decompiled an
existential, I do it now just if `fast_fail:true`.

There was a problem with `failwith` in case of bad entrypoints, I
resolved it by just not inferring the module as a contract.

Finally, `to_option` set `fast_fail:true`, which is unwanted for the
LSP, so I changed it so we can pass `fast_fail:false`.