[#1751] Clear existential types table and add tests
Problem: We run `Checking.type_var_name_tbl.Exists.clear ()`, which
clear the table for existential types in-declarations. But this can lead
to some non-intuitive types when error-recovery is present. A test with
this is added to prevent a regression.

Solution: Don't clear the table. Add tests.

The second test for entrypoint is actually added posteriorly for a
previous fix, where LIGO would crash with "what?" in the presence of an
incorrect entrypoint.