[Chore] Change the default diagnostics pulling mode
Motivation: the default diagnostics pulling mode in `server.ml` is
`OnDocUpdate` which is very slow since it updates them on each keystroke.
Let's change it to `OnDocumentLinkRequest` since it pulls them after
all keystrokes.

Another motivation is that this is a default mode in `vscode`'s config, so,
we stay consistent.