[#1751] Improve printing of recovered modules
Problem: When a module with a signature is recovered, it can have some
missing fields and show existentials.

Solution: To avoid existentials, implement a small hack, transforming a
type into a type var if the RHS is an existential. For more items, I
return the same thing as the happy path even if the case of a type
mismatch.