[#1751] Refactor the Error_recovery module
Problem: The module doesn't have good naming: we use `'` rather than `_`
to avoid conflicts with keywords, and some names like `try` aren't the
best. Moreover, we have a function to handle `option`, but not `result`.

Solution: Rename things, add a new function to handle `result`s and
change the checker to use it.