[#2084] Run metadata downloads with timeout
Problem: with the current implementation, downloads would block LSP
diagnostics. This is fine in the assumption of all metadata being
reasonably small, however if the developer accidentially inserts a link
downloading which would hang, we don't want LSP to hang too.

Solution: add timeouts on downloads.

Callers not directly related to LSP (`ligo compile storage` command)
still use no timeout. While in LSP timeout is mandatory. Timeout is
configurable.

If a download timeouts, we drop a warning. This should not become
annoying for the user:
* If the issue is indeed with the referred site (sometimes it hangs),
  then probably better to remove `tzip16_compatible` attribute from the
  sources to not cause problems for other developers, and verify that
  storage value separately with `ligo compile storage` command.
* If the issue is in the developer's machine (slow traffic), they can
  disable JSON downloading in config.