[#2143] Some optimizations over `references`
Problem: `Def.Hierarchy.create` is very slow. This happens because it
calls `Path.from_absolute` which is a very expensive function (due to
`Path.normalise` being very expensive). On my machine, it would take
around 5 seconds to call `Def.filter_map` on a function that calls that
on over 800000 definitions. With the optimization here, this is reduced
from between 0.14 to 0.24 seconds.

Solution: I did a little optimization over `fold_definitions` that
boosted around 0.1 to 0.2 seconds, but the real optimization was done by
`Files.with_normalized_files`, which is now used by
`Def.Hierarchy.create` and inlay hints. It remembers files and their
`Path`s, so that it's not necessary to re-normalize the same files over
and over.