[Chore] Better handle errors in following passes
Problem: When we call `following_passes_diagnostics`, we call `try`
inside `try_with_lwt`, which is a bad idea, because that `try` will
catch `Local_lwt` errors and log them as `Local_lwt(_)`, which is very
non-informative. We need to switch this, so we have the `try_with_lwt`
inside the `try`.

Solution: Do the inversion. Also, I refactored the `diagnostics` module
a little.

With this, we now see the proper `failwith` messages being logged.