[#2143] Run aggregation pass only once in following passes
Problem: aggregation pass is slow and we run it in the following passes
`n + 1` times where `n` is the number of module entrypoints. We should
reduce this number somehow.

Solution: generate an entrypoint that refers to the all module entrypoints.
It reduced the number of cycles in `following_passes_diagnostics` from
`5.7G` to `4.1G`. For more information read the comments.