[#2183] Set LSP as an environment for analytics
Problem: During analytics collection, when calculating the `env`, we'll
see `(current_process_is_in_docker, is_tty) = (false, false)`, and the
env will be `Ci`. However, this will prevent analytics collection or
proposing analytics to the user.

Solution: Make `env` a thunk and make it check if we are running in LSP
with `is_running_lsp ()`, which is set in `cli.ml` with
`set_is_running_lsp` true`. Now, the `is_in_ci` check will not prevent
running analytics.

I noticed a copy-paste of some code in `cli.ml` to spawn the language
server and refactored it.