[#2183] Better handling of crashes
Problem: When we crash, we still get some notification spam. This will
be bad when we'll be dealing with crash metrics.

Solution: We assume that a crash was resolved if `on_doc` was succesful.
We record the previous exception in order to not show a new
notitification if it's the same as the previous one.