[#2185] Handle pattern anomalies errors
Problem: Pattern anomalies has many usages of unsafe functions (`*_exn`)
which can be reached in case error recovery is enabled and certain
criteria are met. We'd like to better handle those.

Solution: Refactor pattern anomalies to use `raise` rather than unsafe
exceptions. In elaboration, we check anomalies while keeping in mind
that `raise.error` can throw, and in case typer error recovery is
enabled, we log those errors instead.

One thing I did was to refactor `unbound_label_case` to handle records
and sums, rather than assume a record, and to operate on `Ast_typed.row`
rather than `Type.row`. This allowed this exception to be reused from
pattern anomalies, but had a nasty side effect of requiring the only
`Type.row` usage to be decoded. This in turn required to write
`lift_elab` (which I hope this naïve implementation is fine!) and expose
`decode_row`. However, `decode_row` can throw `cannot_decode_texists`,
and so I made `error` in `get_type_of_ctor` be a thunk, so we only
decode the row in case of an error.