[#2173] Fix inlay hints for functions
Problem: Inlay hints for functions may show hints in
the wrong position, if the return type is provided
but not all the parameter types. Reason is that
when return type is provided the function definition
doesn't treated as one needed inlay hint, so variables
in its definition were treated as regular resolved
variables.

Solution: Ban inlay hint for function variable in case
it's return type provided.