[#2187] Support `import` decls up to Ast_core
Problem: In JsLIGO, we currently provide an import declaration. But we currently have a nanopass that simply rejects programs containing these declarations.
In the context of the new build system, we would like to support these
declarations which requires support in the type checker.

Solution: Modify Ast_core to contain all 3 kinds of import declarations.