[Fix] The grammar says:
statement: $ => choice(
  ...,
  $.expression_statement,
  $.declaration,
  ...),

expression_statement: $ => seq($._expressions, $._semicolon),
_expressions: $ => choice($.expression, $.sequence_expression),

expression: ($, previous) => {
  ...,
  $.internal_module,
  ...),

declaration: ($, previous) => choice(
  ...,
  prec('declaration', $.internal_module),
  ...),

When recognising an internal module (a.k.a. namespace), the parser
will chose and expression over a statement, due to implicit precedence
of tree-sitter. The purpose of this commit is to avoid that: wherever
a statement is expected, and a namespace occurs, it will be
interpreted as a declaration (as a statement), not an expression
statement.